fix(release): seal consolidation authority facade
